fix: Update backend preview for compatibility with v14 breaking change
See: https://docs.typo3.org/permalink/changelog:breaking-92434-1761644184
