+++ .github/upcoming/31.md
@@ -4,6 +4,9 @@
 
 - [POTENTIALLY BREAKING] Remove `fluidtypo3/vhs` dependency - check local templates for `v:` fluid helpers (#31)
 
+#### Fix
+- Update backend preview for compatibility with v14 breaking change (see: https://docs.typo3.org/permalink/changelog:breaking-92434-1761644184)
+
 #### Refactor
 
 - Create method for shared variables
